+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # YOUTUBEUNBLOCKUI Changelog
 
+## [1.1.0] - 2025-02-02
+
+> _Important: Please clear your browser cache (e.g.**ctrl+f5**) to ensure that outdated files are updated._
+
+- ADDED: remount menu on startup
+- ADDED: Install and Uninstall
+- FIXED: Layout changes
+
 ## [1.0.0] - 2025-02-02
 
 > _Important: Please clear your browser cache (e.g.**ctrl+f5**) to ensure that outdated files are updated._